Working scheduled functions running in event loop
Current code flashes backlight based on HID report from host.

+++ quantum/scheduled.h
@@ -5,7 +5,7 @@ typedef void (*scheduledFunc)(void);
 typedef struct {
     scheduledFunc func;
     uint16_t timeout;
-    uint16_t times;
+    uint8_t times;
 } ScheduledConsoleItem;
 
 void scheduleFunction(ScheduledConsoleItem item);
